api/requirements.txt
fastapi[all]==0.128.0
uvicorn==0.40.0
SQLAlchemy[asyncio]==2.0.46
asyncpg==0.31.0
psycopg2-binary==2.9.11
python-multipart==0.0.22
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.5